allocation rewrite, delay fixup
this commit fixes a bunch of things, and leaves some things to
be desired. first off, the new allocator rewrite "works", but
the linear allocator does not seem to work on the teensy for
some reason. I will have to look into this, but malloc works
for now. I just know that when I start integrating the SD card,
I will maybe want to have access to a different allocator. or
maybe not.

in addition, delay now takes a sample count argument for its
buffer, which allows it to work on the teensy (which has a
limited memory). yay!

+++ src/extra/linalloc.h
@@ -3,8 +3,8 @@
 
 typedef struct linear_allocator_t {
   void*    base;
-  uint64_t current_offset;
-  uint64_t size;
+  uint32_t current_offset;
+  uint32_t size;
 
 } linear_allocator_t;
